Reduced disk usage when extracting plugin files.
In an attempt to reduce the amount of time my CPU is hogged by the
client tracker, I've added some hashing checks to avoid redundant
directory reconstructions for the plugin files.

Whether this'll actually improve things overall is still up in the air,
but I'll be paying attention to it.